// src/models/selling.ts
/**
 * Selling model.
 */
export class Selling {
    
      id: string;
      
      productId: string;

      productName: string;
    
      productType: string;

      quantity: number;

      price: number;

      quantityForPrice: number;

      buyingPrice: number;

      revenue: number;

      percentage: number;

      total: number;

      time: string;

      createdAt: Date;
}